视频巡查界面视频播放,调试ok;巡查项复合界面创建,调试

+++ OpticalValleyUnite/Classes/UI/VideoPatrol/Model/YQVideoMapPointModel.swift
@@ -11,7 +11,7 @@ import UIKit
 class YQVideoMapPointModel: NSObject {
     
     //定位点id
-    var insPointId : Int64 = 0
+    var insPointId : Int = 0
     
     //定位点type(室内,室外类型  1,室内 2,室外)
     var type : Int = 0
